ファイルの処理をFileUploaderに分割
